chore(Comix): Move api class to network.ts
